Silence warnings

+++ grayScale.c
@@ -3,15 +3,14 @@
 //
 #include "bmplib.h"
 #include <math.h>
-#include<string.h>
 
 PRIVATE int getGray(byte red, byte green, byte blue);
 
 PUBLIC void applyGrayscale(char *imageFileName) {
     byte *data = readImage(imageFileName);
-    int sizeOfFile = getLongFrom4Bytes(&data[2]);
-    int sizeOfHeaders = sizeOfFile - getLongFrom4Bytes(&data[34]);
-    for (int i = sizeOfHeaders; i + 2 < sizeOfFile; i += 3) {
+    unsigned int sizeOfFile = getLongFrom4Bytes(&data[2]);
+    unsigned int sizeOfHeaders = sizeOfFile - getLongFrom4Bytes(&data[34]);
+    for (unsigned int i = sizeOfHeaders; i + 2 < sizeOfFile; i += 3) {
         int grayValue = getGray(data[i], data[i + 1], data[i + 2]);
         data[i] = grayValue;
         data[i + 1] = grayValue;
@@ -22,6 +21,6 @@ PUBLIC void applyGrayscale(char *imageFileName) {
 }
 
 PRIVATE int getGray(byte red, byte green, byte blue) {
-    return round(0.299 * red + 0.587 * green + 0.114 * blue);
+    return (int)round(0.299 * red + 0.587 * green + 0.114 * blue);
 }
